spilt file without main.go
