+++ Backend/database/database.go
@@ -1,0 +1,29 @@
+package database
+
+import (
+	"fmt"
+	"log"
+
+	"gorm.io/driver/postgres"
+	"gorm.io/gorm"
+)
+
+var DB *gorm.DB
+
+const (
+	host = "postgres"
+	// host     = "localhost"
+	username = "myuser"
+	password = "mypassword"
+	dbname   = "mydatabase"
+	port     = 5432
+)
+
+func ConnectDatabase() {
+	dsn := fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=%s sslmode=disable", host, port, username, password, dbname)
+	var err error
+	DB, err = gorm.Open(postgres.Open(dsn), &gorm.Config{})
+	if err != nil {
+		log.Fatalf("failed to connect to database: %v\n", err)
+	}
+}
